api-flask/requirements.txt
flask
flask-cors
Flask-RESTful
Werkzeug
librosa
numpy
docker
PyJWT
flask-pymongo 
python-dotenv
pydub
soundfile
flask_jwt_extended
marshmallow 
